🚚 Rename: features 도메인 구조에 맞게 폴더 이동
