+++ src/shared/ui/card/ReservationCardInfo.tsx
@@ -1,4 +1,4 @@
-import { ReservationCardProps } from '@/features/reservation/ui/reservation-card/ReservationCard';
+import { ReservationCardProps } from '@/features/reservation/my-reservations/ui/ReservationCard';
 import BaseCardInfo from '@/shared/ui/card/base/BaseCardInfo';
 import CardActions from '@/shared/ui/card/CardActions';
 import { cardInfoStyles } from '@/shared/ui/card/cardStyles';
